requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
greenlet==3.3.2
hub-sdk==0.0.24
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.0
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.19.3
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging>=23.0
pandas==2.2.0
pillow>=7.1.2,<11
playwright==1.58.0
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
py-cpuinfo==9.0.0
pyee==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
scipy>=1.11.4
seaborn==0.13.2
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.2.0
torchvision==0.17.0
tqdm==4.67.3
triton==2.2.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.1.0
ultralytics-thop==2.0.18
urllib3==2.6.3
filterpy==1.4.5
streamlit==1.32.0
